+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [4.0.0] - TBD
 
 ### Added
+- Support for running Hive Metastore on Kubernetes.
 - Upgrade to Terraform version 0.12.
 - Configuration variable for `apiary_extensions_version`.
 - Variable to grant cross account AWS IAM roles write access to Apiary managed S3 buckets using assume policy.
@@ -17,6 +18,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Tag VPC endpoint services.
 - Add ansible handler to restart hive metastore services on changes to hive-site.xml and hive-env.sh.
 - add TABLE_PARAM_FILTER environment variable to hive-env.sh on EC2 to fix beekeeper.
+
+### Removed
+- Support for running Hive Metastore on EC2 nodes.
 
 ## [3.0.1] - 2019-08-08
 ### Added